move dagutils package to top level in preparation for merkledag extraction
License: MIT
