+++ dagutils/diffenum.go
