+++ dagutils/diffenum_test.go
